requirements.txt
requests==2.30.0
pydantic==1.10.7
pytest==7.3.1
urllib3==1.26.6
allure-pytest==2.13.5
allure-python-commons==2.13.5